Share RateLimiterStateChip across poke credits tables
