+++ front/components/poke/credits/PokeMembersUsageTable.tsx
@@ -2,13 +2,11 @@ import { AlertChip } from "@app/components/poke/credits/AlertChip";
 import { CreditStateLogsLink } from "@app/components/poke/credits/CreditStateLogsLink";
 import { GrantFreeCreditsButton } from "@app/components/poke/credits/GrantFreeCreditsButton";
 import { MemberConsumptionExportButton } from "@app/components/poke/credits/MemberConsumptionExportButton";
+import { RateLimiterStateChip } from "@app/components/poke/credits/RateLimiterStateChip";
 import { ReconcileCreditStateButton } from "@app/components/poke/credits/ReconcileCreditStateButton";
 import { ResetFairUseButton } from "@app/components/poke/credits/ResetFairUseButton";
 import { PokeDataTable } from "@app/components/poke/shadcn/ui/data_table";
-import type {
-  MemberUsageType,
-  RateLimiterState,
-} from "@app/lib/api/credits/members_usage";
+import type { MemberUsageType } from "@app/lib/api/credits/members_usage";
 import { formatCredits, formatCreditsPrecise } from "@app/lib/client/credits";
 import type { MetronomeAlertRef } from "@app/lib/metronome/alerts/types";
 import { getMetronomeAlertUrl } from "@app/lib/metronome/urls";
@@ -135,17 +133,6 @@ const USER_CREDIT_STATE_CHIP_COLOR: Record<
   on_pool: "success",
   on_pool_low_balance: "warning",
   capped: "warning",
-};
-
-// The rate-limiter's verdict, rendered as a chip. Labels distinguish capped vs
-// near-limit (both warning-toned).
-const RATE_LIMITER_STATE_CHIP: Record<
-  RateLimiterState,
-  { color: "success" | "warning"; label: string }
-> = {
-  capped: { color: "warning", label: "capped" },
-  near_limit: { color: "warning", label: "near limit" },
-  ok: { color: "success", label: "ok" },
 };
 
 // Free seats hold a per-user credit with two balance alerts: "low" (≤20%) and
@@ -415,12 +402,11 @@ function makeColumns({
       header: "Rate limiter state",
       enableSorting: false,
       cell: ({ row }) => {
-        const { rateLimiterState } = row.original;
-        if (rateLimiterState === null) {
-          return <span>—</span>;
-        }
-        const { color, label } = RATE_LIMITER_STATE_CHIP[rateLimiterState];
-        return <Chip size="xs" color={color} label={label} />;
+        return (
+          <RateLimiterStateChip
+            rateLimiterState={row.original.rateLimiterState}
+          />
+        );
       },
     },
     {
