+++ front/lib/api/keys/spend_limit.ts
@@ -5,7 +5,6 @@ import {
   reconcileWorkspaceApiKeyCreditStates,
 } from "@app/lib/api/metronome/reconcile_credit_state";
 import type { Authenticator } from "@app/lib/auth";
-import { getFeatureFlags } from "@app/lib/auth";
 import { roundCreditsToMicroCredits } from "@app/lib/credits/units";
 import {
   clearMetronomeApiKeyCapAlert,
@@ -18,7 +17,6 @@ import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import type { FixedWindowBounds } from "@app/lib/utils/rate_limiter";
 import {
   addFixedWindowCount,
-  getFixedWindowCount,
   readFixedWindowCountWithLazySeed,
 } from "@app/lib/utils/rate_limiter";
 import logger from "@app/logger/logger";
@@ -28,7 +26,6 @@ import type {
   SetApiKeySpendLimitResponse,
 } from "@app/types/api/keys/spend_limit";
 import { isCreditPricedPlan } from "@app/types/plan";
-import type { ModelId } from "@app/types/shared/model_id";
 import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { assertNever } from "@app/types/shared/utils/assert_never";
@@ -424,63 +421,6 @@ export async function isApiKeySpendLimitRateCapReached(
   // The counter stores microCredits; scale the credit threshold up so the
   // comparison stays integer-on-integer.
   return count >= roundCreditsToMicroCredits(threshold);
-}
-
-/**
- * The flag-aware per-key "capped" verdict for each of `keys`, keyed by key model
- * id — the signal the keys UI uses to show the "capped" status. Mirrors the
- * enforcement switch in `isApiKeyBlocked` (lib/api/credits/access_control.ts):
- * with the `enforce_user_spend_limit_rate_cap` flag on, the persisted Metronome
- * `creditState` must not be read — the per-key Redis fixed-window counter
- * (compared to the key's cap) is authoritative; with it off, from
- * `creditState === "capped"`. Fails open (not capped) when the flag is on but
- * the billing cycle can't be resolved or a counter read errors.
- *
- * Lives here (not on `KeyResource`) because the flag read pulls in `lib/auth`,
- * which imports `KeyResource` — computing this inside the resource would be an
- * import cycle. The keys endpoints call this and pass the map to
- * `KeyResource.toJSONWithSpaces`.
- */
-export async function getApiKeysSpendCappedByModelId(
-  auth: Authenticator,
-  keys: KeyResource[]
-): Promise<Map<ModelId, boolean>> {
-  const featureFlags = await getFeatureFlags(auth);
-  const spendCapEnabled = featureFlags.includes(
-    "enforce_user_spend_limit_rate_cap"
-  );
-  if (!spendCapEnabled) {
-    return new Map(keys.map((key) => [key.id, key.creditState === "capped"]));
-  }
-
-  const workspace = auth.getNonNullableWorkspace();
-  const bounds = await resolveSpendLimitCycleBounds(workspace);
-  if (!bounds) {
-    // No resolvable billing cycle: nothing to enforce against (fail-open).
-    return new Map(keys.map((key) => [key.id, false]));
-  }
-
-  // Redis reads (not database queries), one per key of the workspace.
-  const entries = await concurrentExecutor(
-    keys,
-    async (key) => {
-      const cap = key.monthlyCapAwuCredits;
-      if (cap === null) {
-        return [key.id, false] as const;
-      }
-      const result = await getFixedWindowCount({
-        key: makeApiKeySpendLimitAwuCreditsRateLimitKey(key.id),
-        bounds,
-      });
-      // Compare in microCredits to match enforcement exactly
-      // (isApiKeySpendLimitRateCapReached).
-      const capped =
-        result.isOk() && result.value >= roundCreditsToMicroCredits(cap);
-      return [key.id, capped] as const;
-    },
-    { concurrency: 8 }
-  );
-  return new Map(entries);
 }
 
 /**
